+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Description: JAXON
 - Description: Kinova Gen3
 - Description: NEXTAGE